Configure JSON deserialization to accept unknown properties
This configuration is necessary to support different AlpineBits
versions that may have different properties. For example, the 2018-10
Inventory/Basic action introduces an attribute "RoomType", which is
unknown in 2017-10. An ODH entry written by an 2018-10 client, that
includes the "RoomType" attribute would be unreadable by a 2017-10
client, if the deserialization config would not have been adapted.