Improve resolution algorithm of SystemPropertySource
Before this improvement, the SystemPropertySource tried to resolve an ENV variable
by the given key. This could be problematic for keys that contain dots, because
e.g. Linux does not support ENV variables with dots.

Therefore the improved version now first tries to resolve the ENV variable by the
given key. If that fails, it retries the resolution with an altered key that is
an upper-case and dot-replaced version of the original key. For example the key
"some.key" will be transformed to "SOME_FORM" and a resolution with that key is attempted.